web_langbot/requirements.txt

beautifulsoup4==4.9.0
bs4==0.0.1
Django==3.0.5
djangorestframework==3.11.0
psycopg2==2.8.5
pyTelegramBotAPI==3.6.7
requests==2.23.0
gTTS==2.1.1
langdetect==1.0.8
django-filter
sentry-sdk==0.14.3
Pillow
translate
celery
redis
django_extensions
googletrans
